Create DAO aliases and some optimizations

+++ src/TomorrowDAOServer.Application.Contracts/Governance/Dto/IndexerGovernanceSchemeDto.cs
@@ -21,10 +21,10 @@ public class IndexerGovernanceScheme
     public GovernanceMechanism GovernanceMechanism { get; set; }
     public string GovernanceToken { get; set; }
     public DateTime CreateTime { get; set; }
-    public int MinimalRequiredThreshold { get; set; }
-    public int MinimalVoteThreshold { get; set; }
-    public int MinimalApproveThreshold { get; set; }
-    public int MaximalRejectionThreshold { get; set; }
-    public int MaximalAbstentionThreshold { get; set; }
+    public long MinimalRequiredThreshold { get; set; }
+    public long MinimalVoteThreshold { get; set; }
+    public long MinimalApproveThreshold { get; set; }
+    public long MaximalRejectionThreshold { get; set; }
+    public long MaximalAbstentionThreshold { get; set; }
     public long ProposalThreshold { get; set; }
 }
